pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "release_notes_gen"
version = "0.1.0"
description = "Generate release notes from Jira CSV exports using OpenAI"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
]

[project.scripts]
release-notes-gen = "release_notes_gen.__main__:main"

[tool.setuptools.packages.find]
where = ["."]

